// frontend/src/types/contentBlockResponse.ts
/**
 * Generated by orval v7.16.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { FileResponse } from './fileResponse';
import type { JsonNode } from './jsonNode';

export interface ContentBlockResponse {
  id?: number;
  version?: number;
  createdBy?: number;
  updatedBy?: number;
  createdAt?: string;
  updatedAt?: string;
  file?: FileResponse;
  metadata?: JsonNode;
}